pyproject.toml
[project]
name = "linkedin-profile-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.14",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-google-genai>=2.1.5",
    "langchain-groq>=0.3.4",
    "langchain-tavily>=0.2.5",
    "langgraph>=0.5.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "uvicorn>=0.35.0",
]
